+++ src/Symfony/Component/Security/Core/Authorization/TraceableAccessDecisionManager.php
@@ -87,10 +87,14 @@ class TraceableAccessDecisionManager implements AccessDecisionManagerInterface
 
     public function getStrategy(): string
     {
-        // The $strategy property is misleading because it stores the name of its
-        // method (e.g. 'decideAffirmative') instead of the original strategy name
-        // (e.g. 'affirmative')
-        return null === $this->strategy ? '-' : strtolower(substr($this->strategy, 6));
+        if (null === $this->strategy) {
+            return '-';
+        }
+        if (method_exists($this->strategy, '__toString')) {
+            return (string) $this->strategy;
+        }
+
+        return get_debug_type($this->strategy);
     }
 
     /**