+++ src/Symfony/Component/Security/Core/Tests/Authorization/AccessDecisionManagerTest.php
@@ -11,10 +11,12 @@
 
 namespace Symfony\Component\Security\Core\Tests\Authorization;
 
+use PHPUnit\Framework\Assert;
 use PHPUnit\Framework\TestCase;
 use Symfony\Bridge\PhpUnit\ExpectDeprecationTrait;
 use Symfony\Component\Security\Core\Authentication\Token\TokenInterface;
 use Symfony\Component\Security\Core\Authorization\AccessDecisionManager;
+use Symfony\Component\Security\Core\Authorization\Strategy\AccessDecisionStrategyInterface;
 use Symfony\Component\Security\Core\Authorization\Voter\CacheableVoterInterface;
 use Symfony\Component\Security\Core\Authorization\Voter\VoterInterface;
 
@@ -22,6 +24,9 @@ class AccessDecisionManagerTest extends TestCase
 {
     use ExpectDeprecationTrait;
 
+    /**
+     * @group legacy
+     */
     public function testSetUnsupportedStrategy()
     {
         $this->expectException(\InvalidArgumentException::class);
@@ -29,28 +34,86 @@ class AccessDecisionManagerTest extends TestCase
     }
 
     /**
+     * @group legacy
+     *
      * @dataProvider getStrategyTests
      */
     public function testStrategies($strategy, $voters, $allowIfAllAbstainDecisions, $allowIfEqualGrantedDeniedDecisions, $expected)
     {
         $token = $this->createMock(TokenInterface::class);
+
+        $this->expectDeprecation('Since symfony/security-core 5.4: Passing the access decision strategy as a string is deprecated, pass an instance of "Symfony\Component\Security\Core\Authorization\Strategy\AccessDecisionStrategyInterface" instead.');
         $manager = new AccessDecisionManager($voters, $strategy, $allowIfAllAbstainDecisions, $allowIfEqualGrantedDeniedDecisions);
 
         $this->assertSame($expected, $manager->decide($token, ['ROLE_FOO']));
     }
 
     /**
-     * @dataProvider provideStrategies
+     * @dataProvider provideBadVoterResults
+     *
      * @group legacy
      */
-    public function testDeprecatedVoter($strategy)
+    public function testDeprecatedVoter()
     {
         $token = $this->createMock(TokenInterface::class);
+        $strategy = new class() implements AccessDecisionStrategyInterface {
+            public function decide(\Traversable $results): bool
+            {
+                iterator_to_array($results);
+
+                return true;
+            }
+        };
+
         $manager = new AccessDecisionManager([$this->getVoter(3)], $strategy);
 
-        $this->expectDeprecation('Since symfony/security-core 5.3: Returning "3" in "%s::vote()" is deprecated, return one of "Symfony\Component\Security\Core\Authorization\Voter\VoterInterface" constants: "ACCESS_GRANTED", "ACCESS_DENIED" or "ACCESS_ABSTAIN".');
+        $this->expectDeprecation('Since symfony/security-core 5.3: Returning "%s" in "%s::vote()" is deprecated, return one of "Symfony\Component\Security\Core\Authorization\Voter\VoterInterface" constants: "ACCESS_GRANTED", "ACCESS_DENIED" or "ACCESS_ABSTAIN".');
 
         $manager->decide($token, ['ROLE_FOO']);
+    }
+
+    public function provideBadVoterResults(): array
+    {
+        return [
+            [3],
+            [true],
+        ];
+    }
+
+    public function testVoterCalls()
+    {
+        $token = $this->createMock(TokenInterface::class);
+
+        $voters = [
+            $this->getExpectedVoter(VoterInterface::ACCESS_DENIED),
+            $this->getExpectedVoter(VoterInterface::ACCESS_GRANTED),
+            $this->getUnexpectedVoter(),
+        ];
+
+        $strategy = new class() implements AccessDecisionStrategyInterface {
+            public function decide(\Traversable $results): bool
+            {
+                $i = 0;
+                foreach ($results as $result) {
+                    switch ($i++) {
+                       case 0:
+                           Assert::assertSame(VoterInterface::ACCESS_DENIED, $result);
+
+                           break;
+                       case 1:
+                           Assert::assertSame(VoterInterface::ACCESS_GRANTED, $result);
+
+                           return true;
+                   }
+                }
+
+                return false;
+            }
+        };
+
+        $manager = new AccessDecisionManager($voters, $strategy);
+
+        $this->assertTrue($manager->decide($token, ['ROLE_FOO']));
     }
 
     public function getStrategyTests()
@@ -111,14 +174,6 @@ class AccessDecisionManagerTest extends TestCase
                 $this->getVoter(VoterInterface::ACCESS_ABSTAIN),
             ], true, true, true],
         ];
-    }
-
-    public function provideStrategies()
-    {
-        yield [AccessDecisionManager::STRATEGY_AFFIRMATIVE];
-        yield [AccessDecisionManager::STRATEGY_CONSENSUS];
-        yield [AccessDecisionManager::STRATEGY_UNANIMOUS];
-        yield [AccessDecisionManager::STRATEGY_PRIORITY];
     }
 
     public function testCacheableVoters()
@@ -280,6 +335,24 @@ class AccessDecisionManagerTest extends TestCase
         $voter->expects($this->any())
               ->method('vote')
               ->willReturn($vote);
+
+        return $voter;
+    }
+
+    private function getExpectedVoter(int $vote): VoterInterface
+    {
+        $voter = $this->createMock(VoterInterface::class);
+        $voter->expects($this->once())
+            ->method('vote')
+            ->willReturn($vote);
+
+        return $voter;
+    }
+
+    private function getUnexpectedVoter(): VoterInterface
+    {
+        $voter = $this->createMock(VoterInterface::class);
+        $voter->expects($this->never())->method('vote');
 
         return $voter;
     }